24/03/15 pdbs and pdbs_24_03_15 are equals and pdbs_santiago is the version of santiago that we want to implement in ss_culprits

+++ src/search/pdbs_santiago/match_tree.cc
@@ -55,13 +55,12 @@ void MatchTree::build_recursively(
 
     const vector<pair<int, int> > &regression_preconditions = op.get_regression_preconditions();
     Node *node = *edge_from_parent;
-    if (pre_index == static_cast<int>(regression_preconditions.size())) {
-        // All preconditions have been checked, insert op.
+    if (pre_index == regression_preconditions.size()) { // all preconditions have been checked, insert op
         node->applicable_operators.push_back(&op);
     } else {
         const pair<int, int> &var_val = regression_preconditions[pre_index];
 
-        // Set up node correctly or insert a new node if necessary.
+        // setup node correctly or insert a new node if necessary
         if (node->test_var == -1) { // node is a leaf
             node->test_var = var_val.first;
             int test_var_size = g_variable_domain[pattern[var_val.first]];
@@ -71,8 +70,7 @@ void MatchTree::build_recursively(
                 node->successors[i] = 0;
             }
         } else if (node->test_var > var_val.first) {
-            // The variable to test has been left out: must insert new
-            // node and treat it as the "node".
+            // The variable to test has been left out: must insert new node and treat it as the "node".
             Node *new_node = new Node(var_val.first, g_variable_domain[pattern[var_val.first]]);
             // The new node gets the left out variable as test_var.
             *edge_from_parent = new_node;
@@ -80,16 +78,12 @@ void MatchTree::build_recursively(
             node = new_node; // The new node is now the node of interest.
         }
 
-        // Set up edge to the correct child (for which we want to call
-        // this function recursively).
+        // setup edge to the correct child (for which we want to call this function recursively)
         Node **edge_to_child = 0;
-        if (node->test_var == var_val.first) {
-            // Operator has a precondition on test_var.
+        if (node->test_var == var_val.first) { // operator has a precondition on test_var
             edge_to_child = &node->successors[var_val.second];
             ++pre_index;
-        } else {
-            // Operator doesn't have a precondition on test_var:
-            // follow/create *-edge.
+        } else { // operator doesn't have a precondition on test_var, follow/create *-edge
             assert(node->test_var < var_val.first);
             edge_to_child = &node->star_successor;
         }
